feat: upgrade for inferno 4.x (#9)
* Upgrade minor deps

* Update inferno to V4

* Switch to named imports & Fix inconsistencies

Default imports are removed so where possible use infernos auto-import
mechanism and import the rest named.

- Flag `ClassCompononent` is now `ComponentClass`
- Internal flags starting with `_...` have been renamed to shorthands
starting with `$...` (eg. `_lastInput` to `$LI`)

* Rename event doubleClick to dblClick

Inferno changed this to be compliant with react I guess

* Check for vNode and use that if available

The nodes are now wrapped so if $V is available it should be used.

* Change tests to accomodate changes in structure

Keys representation changed from `.Number` to `$Number` I think

* Major upgrade for Inferno V4

* Add prepare hook for installations from Github

* Fix wrong reference for node type

IDRK why this has changed but the tests seem to pass anyways. Weird

* Bump patch version

- Fix wrong nodeType reference while simulating events

* Fix wrong child representation returned by toTree

After checking if there are multiple children (array of children should stay flat)
check if the child is a wrapper for Text Node.
If it is parse the child text node and wrap it in an array (but only if a valid child is returned).