Adjust adapter for inferno V6
- Small change in enzyme adapter (render no longer returns anything)
- Rewritten to tree completely to make code more readable and less error
prone.

toTree handles the complete logic of deciding what to do with
which component types (it trusts getNodeType to report the correct type).

getNodeType handles corrections for types and uses types of children &
inferno's flags in order to decide.

Adding or configuring types should be as straightforward as teaching the
later how to properly distinguish this type and the former how to create
an appropriate tree representation based on this.
It takes inspiration from the design of the official react 16 adapter from enzyme.